API prod deploys require all frontend e2e to pass
Previously this was an informational step to alert the person releasing that their might be an issue.

We are now in the position where CAS1&3 are in public beta and CAS2 is about to go into private beta. The risk of deploying a breaking API change has increased to a point where we should now make this a required step when deploying to preprod and prod.