Delete two production applications used for testing
We had planned to use realistic referral data and leave these referrals in the service. In practice there were a couple of problems with this which mean we want to remove these:

1. Authorship was from the team
2. These referrals might not receive update from Nacro
3. The referral might be double counted somewhere downstream in reporting

We could delete this manually on the prod command line but want to get this more sensitive change in as code where it can be reviewed.