Moved get booking for premises logic
Previously the logic to get a booking for a given premises was fully defined in the premises controller. Whilst some of this logic is specific to premises (e.g. checking if the booking belongs to the given premises), the majority of the logic is generic when retreiving booking.

Because we’re going to add an endpoint to retrieve a booking by its ID only (i.e. no premise ID) to support new withdrawal logic, the common ‘get booking’ logic needs moving into the BookingService

This commit also adds a small premises summary to the Booking API response. This allows the controller to validate that the premises id is as expected, and the premises name will be of use to the UI in subsequent changes