Paginate CAS2 submissions end point
This adds pagination to `/submissions` whichs lists all submitted
applications.

We are using the existing utility functions and `Page` response already
in the API, and utilised by Cas1/3's
`getAllApprovedPremisesApplications` method. We have set the page
`param` to be optional in order not to break the existing frontend - if
no page param is given, then all the results are returned, which means
we will not break the existing UX on the frontend.