Allow tasks types to be defined when getting tasks
TaskService#getAll now accepts a list of task types to be returned.

GetReallocatable controller function now uses this single method instead of individual methods for each tasks types.

This may be slightly less performant when retrieving tasks of a specific type. If we find this to be a problem we can use ‘type specific’ queries when only one type is specified, which will be very easy to introduce now all calls go through the same functions