return `statusDetails` on CAS2 Application Statuses reference data
We previously introduced a set of CAS2 specific
Application statuses, that are applied by
Assessors. We now need to introduce a sub-level of
category on those statuses, which we are calling
'status details'.

This returns the status details under their
corresponding statuses on the reference data
controller via the `ApplicationStatusTransformer`,
so that the frontend can request them and display
as options in a form, and so we have a source for
which Detail belongs to which Status.