POST to create a StatusUpdateDetail on a StatusUpdate
When a user adds a Status and Status Detail(s),
add this information as a Status Update, with a list of Status Update Details
on the Application.

This adds a field `newStatusDetails` to the `Cas2ApplicationStatusUpdate` which
is the data POSTed to endpoint `/submissions/{applicationId}/status-updates`

The `StatusUpdateService` then takes this data and finds the persisted list of statuses/details,
and saves a new `StatusUpdate` and it's optional corresponding `StatusUpdateDetails` in the database.