Hide withdrawable bookings from non CRU users
This adds a configurable `showAllBookings` option to the
`allWithdrawables` service function. If true, we use the
`UserAccessService` to see if the user can manage the booking.
If they can’t we filter them out.

In reality, `currentUserCanManagePremisesBookings` for CAS1 only
looks at a user’s role, so this will only return true for
CRU or AP managers.

When using the withdrawAllForApplication, we set `showAllBookings` to
true, because it doesn’t make sense to withdraw an application and
not withdraw the associated bookings. This may change as the service
evolves, but I think this makes sense as things currently stand.