/tasks now internally directed to /tasks/reallocated
This change reduces duplication of logic, in preparation for adding more logic.

Existing /tasks integration tests had to be updated to set a ‘submittedAt’ date on placement applications as the ‘new’ logic only returns placement applications that have been submitted. This is logical because a placement applciation isn’t considered a ‘task’ until submitted.