Add ‘application’ type to /withdrawables response
An additional type has been added to the list of types that can be returned by the /withdrawables response. This is used to indicate if the application as a whole can be withdrawn by the calling user.

The purpose of this change is to move the logic to determine if an application is withdrawable into the API and out of the UI, simplifying future changes to this logic (which potentially would have to be in the UI and API without this change)