CAS1 Booking Cancellation is now idempotent
In preparation for adding more cascading behaviour to cancellations, the CAS1 cancellation logic is now idempotent (i.e. instead of returning an error if a booking is already cancelled, just return the previously created cancellation entity).

This will allow for simpler cascading logic as we will not need to track which bookings have/have not been cancelled by via other cascading actions

fix