Cascade placement request withdraw to booking
When a placement request is withdrawn, any associated booking is also automatically cancelled. If the booking is already cancelled the call will still succeeed as bookingService#cancel method is now idempotent

If the booking cancellation reutrns an error we’re just logging this (i.e. we don’t propagate this back to the caller). This is because this shouldn’t really happen as the reasons for failure are related to the cancellation reason being invalid, which is hard-coded anyway (and issues will be picked up before deploy by automated tests).