Create cas2_application_notes
we would like CAS2 users to add notes to submitted
applications. we are introducing this in parts -
initially enabling the ability for nomis users to
create notes on applications they submitted

this commit creates the table for the notes, and
introduces a foreign key for the applications and
nomis user tables. In the near future this table
will have a new column for external user id.

This also creates the corresponding Entities to
map to the data.