Applciation withdrawals now relies on cascading
Before this commit when withdrawing an application it would ‘collect’ all withdrawable elements (placement request, placement application and bookings) and withdraw them directly.

In it’s current form this has a few issues:

* These ‘withdrawable’ elements are limited to those withdrawable according to user roles, but in this scenario the user only needs the ability to withdraw the application itself (e.g. an applicant can’t withdraw a booking directly, but that can if that withdrawal cascades from an ancestor element like an applciation or placement request)
* Because withdrawing an entity of one type automatically cascades to sub-types (e.g. placement request withdraws bookings), the top level withdrawal code would call ‘withdraw’ multiple same times for the same elements (i.e. it’s not tracking what elements are being withdrawn via cascading). This this isn’t a functional issue (withdrawals are idempotent), it does make the code harder to comprehend

With this commit, when withdrawing an applciation it will iterate through relates placement requests and placement applications and withdraw them if they’re in a withdrawable state. Withdrawing either of these elements may cascade to bookings automatically.