Added WithdrawalContext to all withdrawal funcs
This is required for entity-specific withdrawal to be able to determine which withdrawal reason when withdrawals are cascaded. Without this when a cascade traverses multiple entities e.g. app -> placement app -> placement request -> booking, the root cause would be lost by the time the call to withdraw the booking is made.

This commit also updates exisiting placement request and placement application logic to use the withdrawal context to determine the withdrawal reason