Use withdrawing entity for PR/PA withdraw reason
If a Placement Request or Placement Applciation is withdrawn as a consequence of some ancestor entity being withdrawn, we should automatically set the withdrawal reason to reflect this. We determine if this is the case by looking at the entity that triggered the withdrawal, made available via the withdrawal context.