CAS1/3 booking cancellation funcs called directly
Previously a generic booking cancellation fucntion was called regardless of the calling CAS. This function would then internally redirected to a CAS-specific cancellation function. We need to make changes to argument nullability for CAS1 to support cascading cancellation reasons which would adversely affect the CAS3 cancellation contract. For this reason we’re now exposing and calling the CAS-speicific functions directly