PR/PA Withdrawal errors for invalid trigger entity
When a Placement Request or Placement Application is withdrawn automatically due to an ancestor entity being withdrawn, we automatically set the withdrawal reason based upon the triggering entity type.

If the triggering entity types is unexpected (i.e. logically shouldn’t happen), we now throw an exception. This should never occur under normal operation but is useful as it allows for completeness in ‘when’ statements and helps catch issues during code changes.

In this change we’ve also renamed ‘reason’ to ‘userProvidedReason’ as the reason is only used if the user is the one directly withdrawing the entity