Simplify PR withdrawal permission check logic
Before this commit the PlacementRequestsService#withdrawPlacementRequest function had a Boolean argument ‘checkUserPermissions’ to determine if user permissions should be checked. This was ‘true’ if the entity was being directly withdrawn or false if indirectly withdrawn (via cascade). This commit determines if the check is required by inspecting the WithdrawalContext, removing the need for this argument