add a CAS2 TimelineEvents Transformer
In order to show several different types of events
in a timeline component on the CAS2 UI, we are
creating a standard shape for them on the
presentation/transformation layer of the API. We
currently need to show - notes, status updates,
and when the application was submitted.

We will follow an existing pattern set by CAS1 for
TimelineEvents component, but as our data has a
different shape and as we are already namespacing
CAS2 work, here we create a new component.

We will shortly return the events on an
Application.