return CAS2 Timeline Events on Submitted Applications
use the previously created transformer to return
TimelineEvents on a submitted application, for
assessors/admins to see.

In order to show several different types of events
in a timeline component on the CAS2 UI, we are
creating a standard shape for them on the
presentation/transformation layer of the API. We
currently need to show - notes, status updates,
and when the application was submitted.