return CAS2 Timeline Events on Applications
use the previously created transformer to return
TimelineEvents on an application, for referrers to
see on the application overview page for
applications they have submitted.

In order to show several different types of events
in a timeline component on the CAS2 UI, we are
creating a standard shape for them on the
presentation/transformation layer of the API. We
currently need to show - notes, status updates,
and when the application was submitted.

I've had to return the events on both the
/applications and /submissions end points, even
though these events will only appear on submitted
applications currently. This is because we have a
division between 'referrer' and 'assessor'
journeys. The /submissions endpoint is queried by
the frontend to retrieve all submitted
applications and is only viewable by CAS2_Admin
and CAS2_Assessor user roles.

The /applications endpoint is queried by the
frontend to retrieve all applications regardless
of whether they have been submitted, and is used
currently by only POM user roles. The frontend
doesn't distinguish when querying for a specific
application whether it is submitted or not - it
just gets /applications/{applicationId}.

I feel like this could be tidied up a bit, but
would need work on the frontend as well/a bit of
architectural thinking.