create sealed class for Cas2User
With the addition of a Note on an Application,
which can be created by either a nomis or external
user, we need a sort of join class to refer to
CAS2 Users. This uses a sealed interface and
extends the `Nomis` and `External` User Entities
so that we can use `when` functions to check what
type of User it is
https://kotlinlang.org/docs/sealed-classes.html#sealed-classes-and-when-expression

We have to keep `email` as a nullable field
because the `NomisDetail` returned from the
`NomisUserRoles` API has email as nullable - we
will return a string saying "Not found", following
previous patterns elsewhere, until we can do
further work to see if this is an issue with our
users, or if most active users will have an email
address.