allow Assessors to create Notes
We previously introduced Notes on an Application
 that could be added a by a Nomis user (e.g. a
Referrer/POM) [1].

This commit allows users with the CAS2_ASSESSOR
role to add notes to any submitted applications.

This is the first time we've come across an entity
that could have either an `ExternalUser` or a
`NomisUser` belonging to it - we discussed
introducing polymorphism in the previous PR, but
tabled it at the time... here, rather than
representing the polymorphism on the database
level, I've introduced a Sealed Class[2] for a
CAS2User which gives us a couple of benefits:

* to construct `when`` blocks that the compiler
will warn us about if we leave out any
possibilities.
* add helper methods to the `CAS2NoteEntity` to assign
a user and ensure it always has either a `Nomis` or
`ExternalUser`.

On the database level, there is a
constraint on the database that a note has either
a `nomis` or an `external` user ID foreign key.

[1] https://github.com/ministryofjustice/hmpps-approved-premises-api/pull/1470
[2] https://kotlinlang.org/docs/sealed-classes.html