Move CAS1 booking email code into standalone class
In preparation for adding booking withdrawal emails, the logic for sending CAS1 booking emails has been moved into a standalone class. Whilst this was the primary purpose of the change, this refactoring makes the Booking Service more cohesive and makes it easier to change (and specifically, test) email changes in the future.

There were 2 unit tests in BookingServiceTest that were related to how ‘length of stay’ was provided in the email parameters. These have been completely moved from BookingServiceTest to the new Cas1BookingEmailServiceTest class.