Assert on assessment info in cas1 app report test
Reviewing the tests for the CAS1 application and referral reports, we were neglecting to assert on assessment columns. This commit ensures that assertion is applied where relevant, and adds a report row for an application that has an incomplete assessment.

It also removes logic in the assessment reassignment function that stripped nano seconds from the assessment created at date/time. This was causing issues during tests because the assessment entity that had been marked as reassigned had a created at date earlier than the now ‘active’ assessment.

If nanos are causing issues in integration tests we can make use of the ‘’ function in the TimeExtensions.roundNanosToMillisToAccountForLossOfPrecisionInPostgres function instead