Add short notice reason to referral report
To access the required data we are accessing assessment JSON data. Whilst this is against best practice (see APS-380), we are taking this approach in the interest of making this change available asap. APS-380 can be used to manage removal of this logic and instead access the data via ‘first class’ entity properties