create Assessment when CAS2 Application is submitted
I haven't included the unhappy path in the
integration test (e.g. checking that an assessment
was not created if the application wasn't
submitted) as we have enough coverage of the
different branching logic in the Service in unit
tests, and anything further would be unnecessarily
checking that the `@Transactional` flag works.