Add `createdByUser` to Appeal response
As the user who created the appeal will now be considered the arbitrator
of the appeal, we’ll need the entire user object to be returned. I’ve
had to make some tweaks to the Appeal integration tests to match on
JSON paths as converting the JSON response to an `Appeal` object gave me
some Jackson weirdness, but I think they’re easier to grok now anyway.