Only the MatchRequest for app date is withdrawable
When returning a list of withdrawable elements, the only time a Match Request should be returned is when it represents the start date/duration specified on the original application (i.e. the one that is automatically created when the application is assessed).

This provides a method for the applicant to withdraw a request for a booking on the original dates without withdrawing the whole application.

Previously we would also return the match requests created for Reuqests for Placements created after the application had been submitted. This is not required because the user can withdraw the Request for Placements themselves which will cascade down the corresponding Match Requests.

As part of this commit general improvements have been made to the Withdrawals integration test when checking the correct elements are being returned as withdrawable, testing against more complex application structures