Remove otherReason validation on CAS1 Withdrawal
Before this commit the API would ensure that when the user chooses ‘otherReason’ when withdrawing an application that free text was provided explaining why. This is no longer required in the new withdrawal solution as this reason isn’t available and currently it’s blocking withdrawal in production where the free text box has been removed on the UI, but not in the list of withdrawal reason.