Migration to add assessments to any submissions without one
In a previous commit we introduced functionality
to create an Assessment whenever a CAS2
applicaiton is submitted [1]. This will have left
any applications in Production that had already
been submitted without an Assessment model. In
order to allow the frontend to expect all
submitted applications should have an assessment,
and to ensure all our data is consistent, we
introduce a migration here to add an Assessment to
any Application that has been submitted and does
not yet have one.

[1] e479df3f5556005cafb15598aed944668b67353e