Optionally return ‘initial’ request for placement
The /applications/{id}/placement-applications endpoint has been enhanced to add an optional query param ‘includeInitialRequestForPlacement’, which is null/false by default. If set to true, an additional PlacementApplication of type ‘initial’ will be returned representing the request for placement implicit in the application submission (i.e. the one based upon the original applications’ arrival date).

This is actually modelled as a PlacementRequest in the database, so certain elements of this specific PlacementApplication should be used with care, specifically the ‘id’ field, which actually corresponds to a PlacementRequest ID. For more information on this approach, see the documentation on PlacementRequestEntity#isForApplicationsArrivalDate