Add concept of submission groups to Placement Apps
In a subsequent commit we’re going to only allow one set of dates for each placement application entity. If multiple dates are provided on submission, multiple placement application entities will be created.

As part of this change we don’t want to lose the ability to track which placement applications dates originated from the same user request.

This commit introduces a new mandatory column placement_applications.placement_application_submission_group_id to track this grouping