Add support for multiple appeals to 'appealed assessments' rule
An assessment can have multiple appeals, in which case the
`appealRepository.findByApplication` method will fail. The return type
of this method has been changed to return a list to correctly handle
this case.

This will allow the tests for the applications reports to run without
triggering this error when they create applications with multiple
appeals.