Return a list of placement apps on submission
This is in prepraration for a change that will split a submitted placement application into one placment application per arrival date on submission

This is a backwards compatible change because the UI does not currently use this return value.

I’ve introduced a new sealed interface ValidatableActionResultError. Having a common interface for validation errors helps us isolate validation logic