Create a Placement App for each date on submission
Previously, the code would use a single Placement Application for requests with multiple arrival dates. This created issues when managing withdrawals because we have a requirement to be able to withdrawal a single arrival date on a given Placement Application, something that this data model does not support.

To overcome this restriction we now create a Placement Application for each requested arrival date on submission by cloning the ‘baseline’ application. All Placement Applicaitons created, including the baseline, are linked by having the same value in ‘placement_applicaiton_submission_group_id’.

Ideally we would remove the concept of mulitple placement dates in the data model, but we would first have to retrofit exisitng Placement Applications in production to match this new model, which is a non-trivial task, especially whilst those Placement Applications are still active (i.e. being assessed, awaiting placement etc.). Therefore, this code continue to support the Placement Applications with multiple arrival dates, but only for exisitng applications