Use correct schema for `UnknownPerson` mapping
Currently the API maps an unknown person to the `FullPerson` schema,
which is causing some tests to fail as an unknown person does not have
the necessary information to construct a `FullPerson` instance when
deserialising the response.

By using the correct model class, this type of error will be eliminated.