seed Assessments on Submitted Applications
we previously introduced[1] a function to create
an Assessment when an Application is submitted.
This ensures our seeded data matches what we would
expect to see on real submitted applications.

I've used the
`@SuppressWarnings("LongParameterList")`
annotation because I don't see a way of reducing
the parameters without making it less readable.

[1]https://github.com/ministryofjustice/hmpps-approved-premises-api/pull/1514/commits/e479df3f5556005cafb15598aed944668b67353e