Send email for each Placement Req on submission
If a placement application has multiple dates, it is split into a placement application per date on submission.

This change sends an email for each placement application created. It also adds additional personalisation to the emails sent to improve the information included in the email and to better differentiate each placement application