Only submitted placement requests are withdrawable
Previously we were returning unsubmitted placement requests. Because these don’t neccessarily have dates associated with them, an error occurred when rendering the withdrawable on the UI. We should not be returning unsubmitted placement apps as they are not surfaced anywhere else on the UI once the user navigates away from the creation forms.