Withdrawal Tree Spike to enable blocking withdrawl
This spike rewrites withdrawal cascading to be centrally managed, using a tree strucutre of related entities to determine what needs withdrawing when cascading.

This approach will make it simple to block ancestor elements being withdrawn when bookings have arrivals. This method on WithdrawalTreeNode can be used for this purpose:

fun isBlocked(): Boolean = status.blocking || children.any { it.isBlocked() }

This recursively drills down the tree to see if any descendents are blocking (to be tested!)

The tree is also used to get the initial list of withdrawables

There are various TODOs to improve things. It would be good if the top level withdrawal request was orchestrated through the tree too but this may be awkward because each entity has a diffierent type used to define the withdrawal reason. BUT - if we didn’t do this i’m not sure how we’d check if the ‘top level’ entity being withdrawn is blocked or not (although the user shouldn’t be requesting this in the first place as it will be marked as blocked when returning withdrawables, but it would be nice to be consistent just-in-case they somehow manage it)