Remove duplicate seeded applications
For some reason we were trying to create Applications with IDs matching
those already created above. The only difference was the `createdAt`
date. We're not quite sure why these were there, but we're assuming it's
safe to delete them.

This doesn't fix the initial problem we were hoping to solve, however,
that of applications already existing when we're trying to run the seed
job.