Use Tree to determine withdrawable entities
We have a requirement to be able to block entities from being withdrawn if they relate a booking that has arrivals. For example, given the following tree of entities:

application
  -> placement application 1
      -> placement request 1
          -> booking with no arrival
  -> placement app 2
      -> placement request 2
          -> booking with arrival

We should not be able to withdraw the application, placement application 2, placement request 2 and the booking with arrival.

To enable such behaviour we need to move towards a tree repsentation of an application’s entities when making decisions about what is and isn’t withdrawable. This commit is the first part of the change, add logic to build the tree  and using that to determine which entities are directly withdrawable by the given user. The existing integration tests provide us good regression for this functionality.

Subsequent commits will use this tree to plan and execute withdrawal cascading, at which point we can introduce the concept of entities blocking the withdrawal of ancestors in the tree.

Building and using the tree has other benefits, such as:

* Reducing code duplication (especially for cascading) and making withdrawals more consistent across all entity types
* Making it easier to comprehend what will and will not be withdrawn by rendering the tree structure using the toString() function
* Making the approach to withdrawals more flexible for future changes to behaviour