Create Assessment on AutoScripted Applications
These are created separately from those in the
`Cas2ApplicationsSeedJob`, and used for seeding using real
NomisUsers and ExternalUser (in dev and test environments)
whose usernames we don't want expose in a seed file.[1]

This separate type of seeding of Applications was missed in [the
PR](https://github.com/ministryofjustice/hmpps-approved-premises-api/pull/1547) for
adding seeded Assessments, meaning when it was deployed to dev, there
were lots of Applications without Assessments. This was tricky to debug
as we were unaware of this separate way of seeding Applications. We'll
document the difference between the two approaches in a `how-to` file.

[1]: https://github.com/ministryofjustice/hmpps-approved-premises-api/pull/1207