Add CasResult for simplified error handling
As part of refactoring withdrawals we need to be able to reason upon the outcome from functions to determine if they’re a success or error. Several existing withdrawal methods use the AuthorisableActionResult<ValidatableActionResult<PlacementApplicationEntity>> result type which makes it difficult to determine this.

This commit introduces a new CasResult class tree which is an amalgamation of [ValidatableActionResult] and [AuthorisableActionResult] (essentially, it flattens them into a single collection of classes). This new type is then used by ApplicationService#withdrawApprovedPremisesApplication as an example/proof of usage.

In time we should look to refactor CasResult into a similar pattern to kotlin.Result, allowing us to provide useful functions on the returned type such as isSuccess() or isFailure() without having to understand the underlying type hierarchy. This should then allow us to simplify generics handling as errors type shouldn’t need to have any knowledge of the <EntityType> returned for success cases, allowing us to then introduce more generic validation functions