Orchestrate PR withdrawal via withdrawal service
In preparation for adding functionality to block withdrawals of ancestor entities when cascading, we want orchestrate all entity withdrawals from the Withdrawal Service.

This has additional benefits

1. It allows for all permission checking related to withdrawals (e.g. can the user directly withdraw this entity, is this entity actually withdrawable) can be performed in one place, using the Withdrawable Tree to determine this in a standardised manner.
2. It removes duplicate functionality from entity-specific services related to permissions checking and triggering cascade withdrawals

In this change a redundant/incorrect integration test is removed because testing a match request email isn’t sent if not for original application is already tested by integration tests withdrawing request for placements.