Bookings with arrivals blocks ancestor withdrawals
If a booking has an arrival, any related entity (match request, request for placement, application) cannot be withdrawn and are marked as blocked. Such entities will not be returned in the withdrawable list, and any attempt to withdraw one of these entities will result in a 400 error.

As part of this commit the WithdrawalTest has been effectively rewritten has so many existing tests broke. This class should now be more concise and each test has a complete description of the entity tree.

This commit also removes unused functions and related unit tests from the Entity Services used for the old approach for determining what can be withdrawn (e.g. getWithdrawablePlacementRequestsForUser)