Create CAS1 App Domain Event Service
The code to build, save and emit the CAS1 Application Submitted Domain Event has been moved into the new service, which is called from ApplicationService.submitApprovedPremisesApplication.

The corresponding unit test previously embedded in the tests for ApplicationService.submitApprovedPremisesApplication have been moved to test this new class

This change was made in preparation for adding more logic into  ApplicationService.submitApprovedPremisesApplication. Prior to this commit, the unit test for this function was far more difficult to understand and modify. Unfortunately the streamline test is still too ‘busy’, but this is a first good step towards reducing complexity.

The code to build, save and emit the CAS1 Application Submitted Domain Event has been moved into the new service, which is called from ApplicationService.submitApprovedPremisesApplication.

The corresponding unit test previously embedded in the tests for ApplicationService.submitApprovedPremisesApplication have been moved to test this new class

This change was made in preparation for adding more logic into  ApplicationService.submitApprovedPremisesApplication. Prior to this commit, the unit test for this function was far more difficult to understand and modify. Unfortunately the streamline test is still too ‘busy’, but this is a first good step towards reducing complexity.