Add applicant and case manager details to API
For CAS1 we want to capture the applicant’s contact details at point of submission (which may be overriden). Furthermore, if the case manager is not the same as the applicant, we want to capture their details too.

For CAS1 we want to capture the applicant’s contact details at point of submission (which may be overriden). Furthermore, if the case manager is not the same as the applicant, we want to capture their details too.

fix