Add nullable `replyToEmailId` parameter
to `EmailNotificationService`'s `sendEmail` function. This will allow us
to specify a `replyTo` email address for emails sent via GOV.UK Notify.
Currently we're defaulting to the CAS1 email address, which isn't right
for CAS2.

If no `replyToEmailId` is passed, we fallback to the default, CAS1,
email address for an email's `replyTo`.

Co-authored-by: Harriet H-W <harriet.horobin-worley@digital.justice.gov.uk>