Send email to assessor when referrer adds note
Here we add a method to send an email to the
shared Assessors inbox if the note was not left by
an external user (which would mean it was left by
an internal HMPPS staff member, i.e. the POM).

Also introducing some utility functions to get
placeholder copy, which could be put in a utility
or the email service if they are needed elsewhere.