Log error to Sentry if referrer has no email
As per commit `7802d942`:

For the edge case that a Nomis user does not have an email,
we want to alert the team that the email has failed for this note, but
not let an error bubble up to the user.

We found that all of our existing 200+ users have emails associated
with their accounts, so we are hoping that we are unlikely to encounter
this.