Removed PersonUtils.kt and moved extension functions into OffenderService.kt.
Changed conditional if to when in getfullInfoForPersonOrThrow() and included check to return NotFoundProblem when personInfo = PersonInfoResult.Unknown. Without this, a ClassCastException is thrown when this condition occurs.

Made getInfoForPerson() private.

Added inner class GetFullInfoForPersonOrThrow to OffenderServiceTest and added 3 x new unit tests.  Removed GetInfoForPerson tests since they are superseded by tests in GetFullInfoForPersonOrThrow.